CAY-1562 CayenneCase unit tests conversion
DataContextBinaryPKTest - fixes for Derby

git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/trunk@1098971 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/cayenne-jdk1.5-unpublished/src/test/java/org/apache/cayenne/access/DataContextBinaryPKTest.java
@@ -48,8 +48,10 @@ public class DataContextBinaryPKTest extends ServerCase {
 
     @Override
     protected void setUpAfterInjection() throws Exception {
-        dbHelper.deleteAll("BINARY_PK_TEST2");
-        dbHelper.deleteAll("BINARY_PK_TEST1");
+        if (accessStackAdapter.supportsBinaryPK()) {
+            dbHelper.deleteAll("BINARY_PK_TEST2");
+            dbHelper.deleteAll("BINARY_PK_TEST1");
+        }
     }
 
     public void testInsertBinaryPK() throws Exception {
